abstract_replication_strategy: prepare options: warn about replication_factor corner cases
Using the replication_factor option with
NetworkTopologyStrategy can currently lead to surprising,
counter-intuitive results.

When creating a keyspace, the option initially sets
the replication factor on all DCs.

But when used in ALTER KEYSPACE queries, the legacy
`replication_factor` option is applied only to
DCs added after the keyspace was created, while
existing DCs are unaffected, and if no DCs were
added, the option is silently ignored.

This change detects when the `replication_factor`
can be applied safely, and otherwise, it prints a warning.

Fixes scylladb/scylladb#8881

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>