abstract_replication_strategy: prepare options: check replication_factor ambiguity
Using the replication_factor option with
NetworkTopologyStrategy can currently lead to surprising,
counter-intuitive results.

When creating a keyspace, the option initially sets
the replication factor on all DCs.

But when used in ALTER KEYSPACE queries, the legacy
`replication_factor` option is applied only to
DCs added after the keyspace was created, while
existing DCs are unaffected, and if no DCs were
added, the option is silently ignored.

scylladb/scylladb#8881 suggests adding a warning
in this case, but that might not be enough
as practically, looking at our testing code base,
the common assumption is that the `replication_factor`
option would apply in trivial cases, like when there's
only a single data center, or in successive use
after CREATE KEYSPACE, when all data centers continue
to use the same replication factor.

This change detects when the `replication_factor`
can be applied safely, or return an error if
the new `replication_factor` is ambiguous.

Allow applying the replication factor only in the following cases:
- Data centers are divided into 2 groups:
  configured and unconfigured DCs
- If no DC is configured, apply the replication_factor to all DCs.
- Otherwise, if all configured DCs have the same RF:
  - If there are no unconfigured DCs, just apply the new
    replication_factor too all DCs.
  - If there are both configured DCs (all configured the same),
    and one or more unconfigured DCs, apply the replication_factor
    to the unconfigured DCs (like it is done today) iff
    the replication_factor is the same as the one already configured
    to the presently configured DCs
- Otherwise, the replication_factor is deemed as ambiguous
  and an error will be returned.

Fixes scylladb/scylladb#8881

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>