distributed_loader: Use state as _sstable_directories key
The populator maintains a map of path -> sstable_directory pairs one for
each subdirectory for every sstable state. The "path" is in fact not
used by the logic as it's just a subdirectory name for the state and the
rest of the core operates on state. So it's good to make the map of
directories also be indexed by the state.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>