distributed_loader: Move check for the missing dir upper
The quarantine directory can be missing on the datadir and that's OK. In
order to check that and skip population the populator code uses two-step
logic -- first it checks if the directory exists and either puts or not
the sstable_directory object into the map. Later it checks the map and
decide whether to throw or not if the directory is missing.

Let's keep both check and throw in one place for brevity.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>