main: use fallback parameter when converting a YAML node
as yaml-cpp returns an invalid node when the node to be indexed
does not exist all. but it allows us to provide a fallback value
which is returned when the node is not valid. so, let's just use
this helper for accessing a node which does not necessarily exist.

simpler this way

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>