test/object_store: keep fd before closing it
it turns out after closing an fd (1) representing stdout, "1"
will be representing stderr, so the loop which is supposed to
redirect both stdout *and* stderr will end up with redirecting
stderr only. so in order to preserve the logging messages
printed to stdout, we need to keep the fds before closing them.

so, in this change, we keep the file and corresponding fd before
the loop. and use them in the loop.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>