table: compaction_group_for_token: extend verfication ifndef SCYLLA_BUILD_MODE_RELEASE
Specifically verify also the minimum and maximum tokens
that may not be contained in compaction group range.

The minimum token will be used in the next patch
to look up the first compaction group.

Put the verification logic undef
`#ifndef SCYLLA_BUILD_MODE_RELEASE` to disable the
extra verification in production.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>