cql-pytest: Dont flush out/err between fork and exec
When fork+exec-ing a process with stdio redirection it's recommended to
flush stdout/err in-memory buffers not to make child and parent
processes produce buffered message twice. This recommendation applies to
fork() -- this is where buffers are duplicated and have chances to get
pronted twise. Flushing after fork() only makes sense to make sure
anything that was printed before exec() finds its way to the descriptor,
but it's not the case here.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>