commitlog_replayer: refactor commitlog_replayer::impl::init
We don't need map_reduce here since get_truncated_positions returns
the same result on all shards.

We remove 'finally' semantics in this commit since it doesn't seem we
really need it. There is no code that relies on the state of this
data structure in case of exception. An exception will propagate
to scylla_main() and the program will just exit.