doc: fix the descriptions of config parameters
This commit fixes the descriptions of configuration
parameters in the config.cc file in terms of
language correctness, especially typos, punctuation,
grammar, capitalization, and minor wording changes.

This commit DOES NOT alter the parameter descriptions
in a way that may change their meaning (such updades
will be added with follow-up PRs, as they need to be
reviewed by subject matter experts.

Backgroud:
The descriptions of the parameters from the config.cc
file are now published and available in user-facing
documentation, so we need to ensure they are both
correct and clear.